+++ src/index.ts
@@ -1,3 +1,8 @@
+export interface LinkOptions {
+  documents?: boolean;
+  repository?: string;
+}
+
 export interface MarkdownOptions {
   tagFields?: string[];
   manifests?: boolean;
@@ -6,4 +11,5 @@ export interface MarkdownOptions {
   stylesheet?: string;
   root?: string;
   presentation?: string;
+  links?: LinkOptions;
 }